style(command): move into its own class
